Corrected typo in README example
The example of the download dataset had a typo, where it was using the json files of `pen-animals-Jul-28.json` in order to download the `--output-dir /home/data/MineRLBasaltFindCave-v0`. 

I don't know if these commands were used with copy / paste to get the data on the server, but it might be worth checking if the `/home/data/MineRLBasaltFindCave-v0` and `/home/data/MineRLBasaltCreateVillageAnimalPen-v0` are the same.